requirements.txt
streamlit==1.32.2
tensorflow-cpu==2.15.0
numpy==1.24.3
pillow==10.2.0
h5py==3.10.0
gdown==5.1.0